mv: peer state to cluster manager
