+++ src/services/query_io.rs
@@ -1,9 +1,10 @@
 use std::fmt::Write;
 
-use super::cluster::actors::peer::PeerState;
 use crate::services::statefuls::cache::CacheValue;
 use anyhow::{Context, Result};
 use bytes::{Bytes, BytesMut};
+
+use super::cluster::manager::PeerState;
 
 // ! CURRENTLY, only ascii unicode(0-127) is supported
 const FILE_PREFIX: char = '\u{0066}';
